// frontend/src/app/shared/models/servicio-turistico.model.ts
export interface ServicioTuristico {
  idEstab: string;
  subcategoriaServicio?: string;
  nombreServicio?: string;
  tipoServicio?: string;
  contacto?: string;
  linkReserva?: string;
  descripcion?: string;
  vehiculoTipo?: string;
  zonaCobertura?: string;
  horarioServicio?: string;
  lineas?: string;
  frecuencia?: string;
  tipoRamal?: string;
  programasOfrecidos?: string;
  acreditaciones?: string;
  cantidadAsesores?: number;
}